# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library Management API - Use Our API To Have Custom Workflows And Integration!",
        "description": "Library Management API",
        "contact": {
            "email": "admin@mail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://library.test/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/book-circulations": {
            "get": {
                "tags": [
                    "Circulated Books List"
                ],
                "summary": "List of books ",
                "operationId": "getAllIssuedBookList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search for book related info e.g = title,author,genre",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "book name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Error: Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Book Issue"
                ],
                "summary": "Create Issue book",
                "description": "Issue or Return book",
                "operationId": "createIssueBook",
                "parameters": [
                    {
                        "name": "user_name",
                        "in": "query",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "book_name",
                        "in": "query",
                        "description": "list of book id",
                        "required": true,
                        "schema": {
                            "type": "object"
                        },
                        "example": "[1,2,3]"
                    },
                    {
                        "name": "return_date",
                        "in": "query",
                        "description": "book return date",
                        "required": true,
                        "schema": {
                            "type": "date"
                        },
                        "example": "05-11-2022"
                    },
                    {
                        "name": "note",
                        "in": "query",
                        "description": "issue note",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "text note"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "201"
                                        },
                                        "data": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/book-circulations/return-book/{user_id}/submit": {
            "patch": {
                "tags": [
                    "Return Book"
                ],
                "summary": "return an issue book",
                "description": "return issue book",
                "operationId": "returnBook",
                "parameters": [
                    {
                        "name": "book_name",
                        "in": "query",
                        "description": "list of book id",
                        "required": true,
                        "schema": {
                            "type": "object"
                        },
                        "example": "[1,2,3]"
                    },
                    {
                        "name": "note",
                        "in": "query",
                        "description": "return note",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "text note"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "201"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
...
    },
    "security": [
        {
            "passport": []
        }
    ]
}